Change pure virtual matVec interface to use standard library types.
Defined the type vec_ptr that is a shared_ptr to a vector of fp_type elements.
The vec_ptr is now passed in and out of matVec for a memory safe
implementation. An overloaded matVec function accepts raw array pointers,
converts to a vec_ptr and calls the interface.