Issue47 statepattern (#53)
Closes #47: Creates interfaces for subclass methods of SparseMatrix

Description:
Define 3 pure virtual functions on the SparseMatrix base class:
 - assemblyStorage
 - matVec
 - _unAssemble

These definitions make the SparseMatrix class an abstract class
and require that the subclasses implement something. Otherwise
the code won't compile. Presently the test does not compile.

* WIP #47: Create subclass for testing in testSparseMatrix

Description:
The previous commit changed SparseMatrix to an abstract type, so
the unit test does not compile. This commit modifies the unit test
to create a concrete subclass of SparseMatrix so that the test
can compile.

* WIP #47: Adding enumerations for matrix state

Description:
Added enums for the matrix state and attributes.
Not currently used.

* Fixing initialization of size_t data on SparseMatrix

Description:
When initializing a size_t to -1 it's value is translated to
the largest value representable by size_t. It is more correct to
initialize to 0.

* WIP #47: Create skeleton code for setCoefficient

Description:
This commit defines the setCoefficient interface and defines an
empty routine.

* WIP #47: Adjust indentation in sparsemat.cpp

Description:
Trying to stick with 4-spaces for indentation.

* WIP #47: explicitly adding the implicit "this" to methods

Description:
In C++ for member functions on a class, the class is referencable
as the default variable "this" to refer to "this object" on which
the member function was called.

* WIP #47: Implements setCoefficient

Description:
This implements the setCoefficient method by making use of the
standard library's map container that is ordered by the pair
template type.

* WIP #47: Update SparseMatrix constructor/destructor

Description:
Updates the constructor and destructor to utilize the MatrixState
attribute to control logic around the object's state.

* WIP #47 Fixing indentation in setCoefficient

* WIP #47: Change private to protected

Description:
Changing private attributes to protected so subclasses can modify.

* WIP #47: Adding some more content to the test

Description:
Adding some content to the test so illustrate how to implement
the base class.

* Change constructor of SparseMatrix to use member initializer list

* Change pure virtual matVec interface to use standard library types.

Defined the type vec_ptr that is a shared_ptr to a vector of fp_type elements.
The vec_ptr is now passed in and out of matVec for a memory safe
implementation. An overloaded matVec function accepts raw array pointers,
converts to a vec_ptr and calls the interface.

* Change using keyword to typedef for build error

* Enforce C++11 standard for Travis build errors

Co-authored-by: Eric Pearson <peareric@umich.edu>
Co-authored-by: Eric P <38959114+peareric@users.noreply.github.com>