Implemented a skeleton code for csr matrix class (#51)
* Implemented a skeleton code for csr matrix class

* WIP: Add csr skeleton files
This commit adds a .cpp file based on the testSpMat and updates the .hpp
file accordingly. It also attempts to add a testcase to make sure the
skeleton file is compiling correctly.

* Add csr class skeleton with compile test.
Added csr templated class in .hpp file with .cpp file that specifies
a double and a float implementation. It is currently a skeleton code
where each class method prints that it is executing when called.
Added csrTest to the test directory. The test creates a csr object
and calls some methods.
Updated CMakeLists files to compile the .cpp file and csrTest.

* Fixed minor typo in documentation on line 16

* fixed a different typo on line 16

Co-authored-by: Mark Lee <marklee@Marks-MacBook-Pro.local>
Co-authored-by: Kristin Nicole Barton <knbarton@gl-login2.arc-ts.umich.edu>
Co-authored-by: deanrp2 <37961031+deanrp2@users.noreply.github.com>